adding dask utils testing
